fix: pandas 2.2 + pyarrow dtype enforcement bugs (F1-F5) + pyarrow dep
F1 MEDIUM — paper_logger.py log_outcome(): all df.at[] assignments now cast
  to str. _read_df() uses dtype=str so every cell must be a string; raw
  float/int values raised TypeError with pyarrow-backed pandas 2.2.
  _cov() helper changed to return '' / str(int(...)) instead of np.nan / int.

F2 MEDIUM — paper_logger.py _upsert(): df.at[idx, col] = str(row[col]).
  Same root cause — idempotent re-log path also wrote raw floats into
  the string-typed DataFrame, crashing on second signal write for same date.

F3 LOW — tests/test_remediation.py _spx_df(): Volume column created with
  .astype(float) so the test_no_change_no_leakage multiplication by 1.5
  does not attempt to store floats in an int64 column.

F4 MEDIUM — signal_generator.py _StackingEnsemble.__init__():
  Added self.task = 'regression'. ConformalPredictor.__init__ checks
  model.task != 'regression' and raised AttributeError on the fallback
  ensemble, preventing both high and low models from calibrating and
  producing all-zero signals.

F5 LOW — signal_generator.py _StackingEnsemble.fit():
  HuberRegressor max_iter raised from 300 to 1000. Previous value caused
  ConvergenceWarning on every run against the full ~6k-row training set.

requirements.txt — added pyarrow>=14.0. The entire pipeline reads and
  writes Parquet files but pyarrow was not listed, causing
  'ImportError: Unable to find a usable engine' on fresh installs.

Test suite: 380 passed, 6 skipped, 0 failed.

+++ tests/test_remediation.py
@@ -41,7 +41,7 @@ def _spx_df(n: int = N, seed: int = 42) -> pd.DataFrame:
     h = c + rng.uniform(4, 25, n)
     l = c - rng.uniform(4, 25, n)
     o = c + rng.normal(0, 4, n)
-    v = rng.integers(1_000_000, 5_000_000, n)
+    v = rng.integers(1_000_000, 5_000_000, n).astype(float)  # FIX F3: float avoids int*1.5 dtype error
     return pd.DataFrame({"Open": o, "High": h, "Low": l, "Close": c, "Volume": v},
                         index=dates)
